server/requirements.txt
numpy
fastapi==0.104.1
uvicorn==0.24.0.post1
regex
jinja2==3.1.2
python-multipart
bs4
typing
requests
python-dotenv
starlette